by andrewbelcher: Move the createUser method into a trait so we can use it elsewhere.
