+++ tests/src/Kernel/AcquisitionApiTest.php
@@ -8,7 +8,7 @@
 namespace Drupal\Tests\decoupled_auth\Kernel;
 
 use Drupal\KernelTests\KernelTestBase;
-use Drupal\decoupled_auth\Entity\DecoupledAuthUser;
+use Drupal\decoupled_auth\DecoupledAuthUserCreationTrait;
 use Drupal\decoupled_auth\AcquisitionServiceInterface;
 
 /**
@@ -18,6 +18,8 @@ use Drupal\decoupled_auth\AcquisitionServiceInterface;
  * @group decoupled_auth
  */
 class AcquisitionApiTest extends KernelTestBase {
+
+  use DecoupledAuthUserCreationTrait;
 
   /**
    * {@inheritdoc}
@@ -32,42 +34,6 @@ class AcquisitionApiTest extends KernelTestBase {
 
     $this->installEntitySchema('user');
     $this->installSchema('system', 'sequences');
-  }
-
-  /**
-   * Create a user with the given email and name.
-   *
-   * @var string $name
-   *   This is suffixed with '@example.com' for the mail and, if not decoupled,
-   *   is used for the name of the user. If not given, a random name will be
-   *   generated.
-   * @var bool $decoupled
-   *   Whether this should be a decoupled user. Defaults to FALSE.
-   *
-   * @return \Drupal\decoupled_auth\Entity\DecoupledAuthUser
-   *   The created user.
-   */
-  protected function createUser($name = NULL, $decoupled = FALSE) {
-    // Generate a random name if we don't have one.
-    if (!$name) {
-      $name = $this->randomMachineName();
-    }
-
-    // Create and save our user.
-    $user = DecoupledAuthUser::create([
-      'mail' => $name . '@example.com',
-      'name' => $decoupled ? NULL : $name,
-      'status' => 1,
-    ]);
-    $user->save();
-
-    // Set the given name as a property so it can be accessed when the user is
-    // decoupled.
-    $user->original_name = $name;
-
-    $this->assertTrue($user, 'User successfully created.');
-
-    return $user;
   }
 
   /**
